[Skia][GTK][WPE][Win] Extract DragImageSkia https://bugs.webkit.org/show_bug.cgi?id=293657
Reviewed by Don Olmstead and Carlos Garcia Campos.

Extracted Skia implementation of DragImage from DragImageGtk into DragImageSkia
and started using it on Windows and WPE.

* Source/WebCore/PlatformGTK.cmake:
* Source/WebCore/PlatformWPE.cmake:
* Source/WebCore/PlatformWin.cmake:
* Source/WebCore/SourcesGTK.txt:
* Source/WebCore/platform/DragImage.h:
* Source/WebCore/platform/gtk/DragImageGtkCairo.cpp: Copied from Source/WebCore/platform/gtk/DragImageGtk.cpp.
Only Cairo-specific bits are left in the file, so it was renamed.

(WebCore::dragImageSize):
(WebCore::deleteDragImage):
(WebCore::scaleDragImage):
(WebCore::dissolveDragImageToFraction):
(WebCore::createDragImageFromImage):
(WebCore::createDragImageIconForCachedImageFilename):
(WebCore::createDragImageForLink):
(WebCore::createDragImageForColor):
* Source/WebCore/platform/mediastream/libwebrtc/gstreamer/GStreamerVideoEncoderFactory.cpp: Non-unified build fix.

* Source/WebCore/platform/skia/DragImageSkia.cpp: Renamed from Source/WebCore/platform/gtk/DragImageGtk.cpp.
Skia specific code from DragImageGtk, verbatim.

(WebCore::dragImageSize):
(WebCore::deleteDragImage):
(WebCore::scaleDragImage):
(WebCore::dissolveDragImageToFraction):
(WebCore::createDragImageFromImage):
(WebCore::createDragImageIconForCachedImageFilename):
(WebCore::createDragImageForLink):
(WebCore::createDragImageForColor):
* Source/WebCore/platform/win/DragImageWin.cpp:
Merged remaining logic into DragImageWinCairo.cpp as it is all specific to Cairo.

* Source/WebCore/platform/cairo/DragImageCairo.cpp: Copied from Source/WebCore/platform/gtk/DragImageGtk.cpp.
Droped GTK3 specific code and moved the rest of cairo code to cairo directory.

Canonical link: https://commits.webkit.org/295695@main
