+++ Source/WebCore/SourcesGTK.txt
@@ -89,7 +89,6 @@ platform/graphics/x11/XErrorTrapper.cpp @no-unify
 
 platform/gtk/CursorGtk.cpp
 platform/gtk/DragDataGtk.cpp
-platform/gtk/DragImageGtk.cpp
 platform/gtk/GRefPtrGtk.cpp
 platform/gtk/GtkUtilities.cpp @no-unify
 platform/gtk/LocalizedStringsGtk.cpp
