+++ Source/WebCore/platform/cairo/DragImageCairo.cpp
@@ -25,28 +25,14 @@
 #include "TextFlags.h"
 #include "TextIndicator.h"
 #include <cairo.h>
-#include <gdk/gdk.h>
 #include <wtf/URL.h>
-
-#if USE(SKIA)
-WTF_IGNORE_WARNINGS_IN_THIRD_PARTY_CODE_BEGIN
-#include <skia/core/SkBitmap.h>
-#include <skia/core/SkCanvas.h>
-#include <skia/core/SkImageInfo.h>
-WTF_IGNORE_WARNINGS_IN_THIRD_PARTY_CODE_END
-#endif
 
 namespace WebCore {
 
 IntSize dragImageSize(DragImageRef image)
 {
-#if USE(CAIRO)
     if (image)
         return { cairo_image_surface_get_width(image.get()), cairo_image_surface_get_height(image.get()) };
-#elif USE(SKIA)
-    if (image)
-        return { image->width(), image->height() };
-#endif
 
     return { 0, 0 };
 }
@@ -68,7 +54,6 @@ DragImageRef scaleDragImage(DragImageRef image, FloatSize scale)
     if (imageSize == scaledSize)
         return image;
 
-#if USE(CAIRO)
     RefPtr<cairo_surface_t> scaledSurface = adoptRef(cairo_surface_create_similar(image.get(), CAIRO_CONTENT_COLOR_ALPHA, scaledSize.width(), scaledSize.height()));
 
     RefPtr<cairo_t> context = adoptRef(cairo_create(scaledSurface.get()));
@@ -80,20 +65,6 @@ DragImageRef scaleDragImage(DragImageRef image, FloatSize scale)
     cairo_paint(context.get());
 
     return scaledSurface;
-#elif USE(SKIA)
-    auto imageInfo = SkImageInfo::Make(scaledSize.width(), scaledSize.height(), image->imageInfo().colorType(), image->imageInfo().alphaType());
-    SkBitmap bitmap;
-    bitmap.allocPixels(imageInfo);
-
-    SkPixmap pixmap;
-    if (!bitmap.peekPixels(&pixmap))
-        return nullptr;
-
-    if (!image->scalePixels(pixmap, SkSamplingOptions(SkCubicResampler::CatmullRom())))
-        return nullptr;
-
-    return SkImages::RasterFromBitmap(bitmap);
-#endif
 }
 
 DragImageRef dissolveDragImageToFraction(DragImageRef image, float fraction)
@@ -101,35 +72,12 @@ DragImageRef dissolveDragImageToFraction(DragImageRef image, float fraction)
     if (!image)
         return nullptr;
 
-#if !USE(GTK4)
-    if (!gdk_screen_is_composited(gdk_screen_get_default()))
-        return image;
-#endif
-
-#if USE(CAIRO)
     RefPtr<cairo_t> context = adoptRef(cairo_create(image.get()));
     cairo_set_operator(context.get(), CAIRO_OPERATOR_DEST_IN);
     cairo_set_source_rgba(context.get(), 0, 0, 0, fraction);
     cairo_paint(context.get());
 
     return image;
-#elif USE(SKIA)
-    SkBitmap bitmap;
-    bitmap.allocPixels(image->imageInfo());
-
-    SkPixmap pixmap;
-    if (!bitmap.peekPixels(&pixmap))
-        return nullptr;
-
-    auto canvas = SkCanvas::MakeRasterDirect(bitmap.info(), pixmap.writable_addr(), bitmap.rowBytes());
-
-    SkPaint paint;
-    paint.setAlphaf(fraction);
-
-    canvas->drawImage(image, 0, 0,  { }, &paint);
-
-    return SkImages::RasterFromBitmap(bitmap);
-#endif
 }
 
 DragImageRef createDragImageFromImage(Image* image, ImageOrientation, GraphicsClient*, float)
